requirements.txt
anyio==3.6.2
click==8.1.3
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.95.1
h11==0.14.0
httptools==0.5.0
idna==3.4
motor==3.1.2
pika==1.3.2
pydantic==1.10.7
pymongo==4.3.3
python-dotenv==1.0.0
PyYAML==6.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
